refactor: move files from store to modules folder
