+++ src/modules/real-time-metrics/actions/load-dataset-available-aggregations.js
