+++ src/modules/real-time-metrics/actions/load-info-available-filters.js
