+++ src/modules/real-time-metrics/actions/load-pages-dashboards.js
