+++ src/modules/real-time-metrics/actions/load-report-with-meta.js
