+++ src/modules/real-time-metrics/gql/available-aggregations-query.js
