+++ src/modules/real-time-metrics/gql/input-list-by-dataset-query.js
