+++ src/modules/real-time-metrics/helpers/convert-beholder-to-chart.js
