+++ src/modules/real-time-metrics/helpers/convert-report-meta-to-gql.js
