+++ src/modules/real-time-metrics/helpers/extract-filters-and-variables.js
