+++ src/modules/real-time-metrics/helpers/fill-result-query.js
