+++ src/modules/real-time-metrics/helpers/filter-to-graphql-string.js
