+++ src/modules/real-time-metrics/helpers/format-dataset-filter-to-gql.js
