+++ src/modules/real-time-metrics/helpers/format-dataset-variables.js
