+++ src/modules/real-time-metrics/helpers/mount-filter-types.js
