+++ src/modules/real-time-metrics/helpers/validate-filters.js
