+++ src/modules/real-time-metrics/helpers/verify-blacklist-fields.js
