+++ src/modules/real-time-metrics/helpers/verify-whitelist-fields.js
